change the postmapping deleteuser to deletemapping

+++ src/main/java/com/bank/account/jpa/BankAccountJPA/AccountController.java
@@ -5,6 +5,7 @@ import java.util.ArrayList;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -123,10 +124,14 @@ public class AccountController {
 		return fetchDataService.findById(AccNum);
 	}
 	
-	@PostMapping("/DeleteUser/{AccNum}")
+	@DeleteMapping("/DeleteUser/{AccNum}")
 	public void deleteUser (@PathVariable("AccNum") int AccNum)
 	{
-		fetchDataService.deleteById(AccNum);
+		if(fetchDataService.existsById(AccNum))
+		{
+			fetchDataService.deleteById(AccNum);
+		}
+	
 	}
 	
 	@PostMapping("/Withdraw/{id}/{Amount}")
